+++ include/grapho/camera/viewport.h
@@ -5,14 +5,14 @@ namespace grapho {
 namespace camera {
 struct Viewport
 {
-  int Left = 0;
-  int Top = 0;
-  int Width = 0;
-  int Height = 0;
+  float Left = 0;
+  float Top = 0;
+  float Width = 0;
+  float Height = 0;
   std::array<float, 4> Color = { 1, 0, 1, 0 };
   float Depth = 1.0f;
 
-  float AspectRatio() const { return (float)Width / (float)Height; }
+  float AspectRatio() const { return Width / Height; }
 };
 
 } // namespace
